refactor(dashboard): restructure maily editor files and extract variables logic
