+++ apps/dashboard/src/components/workflow-editor/steps/email/email-preview.tsx
@@ -1,5 +1,5 @@
 import { Avatar, AvatarImage } from '@/components/primitives/avatar';
-import { MAILY_EMAIL_WIDTH } from '@/components/workflow-editor/steps/email/maily';
+import { MAILY_EMAIL_WIDTH } from '@/components/workflow-editor/steps/email/maily-config';
 import { cn } from '@/utils/ui';
 import { HTMLAttributes, useEffect, useRef } from 'react';
 import { RiArrowDownSFill } from 'react-icons/ri';
