+++ apps/dashboard/src/components/workflow-editor/steps/email/views/variable-view.tsx
